refining angular map
